Add files via upload
Directory to be used with intelliJ.
Includes simple main class (not connected to gui)
Includes new class MainWindow which holds the main window (duh!) and a scene each for the 4 scenes we implementing.
Each scene is set by a dedicated class. 
class includes a set of methods to set the scene of the window to the desired scene.